+++ site-lisp/projectile/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+### New features
+
+* Add [boot-clj](https://github.com/boot-clj/boot) project type.
+* Add support for projects using gradlew script.
+
 ## 0.13.0 (10/21/2015)
 
 ### New features